backend/requirements.txt
# CoreAstra Backend Dependencies
# AI-Powered Terminal & Intelligent Control Interface

# Web Framework
fastapi==0.124.4
uvicorn==0.38.0
websockets==12.0
python-multipart==0.0.6

# AI Engines
google-generativeai==0.4.0
groq==0.4.2
anthropic==0.18.0
ollama==0.1.6

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.12.5
pydantic-settings==2.1.0

# Database & Storage
sqlalchemy==2.0.37
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
psutil==5.9.8
httpx==0.25.2

# Logging
loguru==0.7.2
